Refactor benchmark example
